+++ src/workflow/helpers.ts
@@ -1,4 +1,8 @@
-import { TemplateProjectType } from "../utils/constants.js";
+import { Selections } from "../types.js";
+import {
+  FullstackBoilerplateTemplateInfo,
+  TemplateProjectType,
+} from "../utils/constants.js";
 
 export const needTemplateChoice = (prev: any) => {
   switch (prev) {
@@ -14,6 +18,14 @@ export const needSigningOptionChoice = (prev: any) => {
     default:
       return null;
   }
+};
+
+export const needSurfChoice = (values: Selections) => {
+  if (values.template.path === FullstackBoilerplateTemplateInfo.value.path) {
+    return "select";
+  }
+
+  return null;
 };
 
 export const needFrameworkChoice = (prev: any) => {