+++ templates/boilerplate-template/frontend/utils/surfClient.ts
@@ -1,0 +1,8 @@
+import { createSurfClient } from "@thalalabs/surf";
+import { aptosClient } from "./aptosClient";
+
+const surf = createSurfClient(aptosClient());
+
+export function surfClient() {
+  return surf;
+}